evi-brush-actions-ready/python/requirements.txt

diffusers
transformers
accelerate
safetensors
torch
xformers; platform_system=="Windows" or platform_system=="Linux"
gradio
pillow
